fix: resolve ~162 Ktor server compile errors (server compiles clean)
Key fixes: unclosed /* comment in ChatRepository (swallowed ~900 lines),
missing Exposed imports (count, or, notInSubQuery, countDistinct extension),
countDistinct as Column extension (not top-level), tables without id column
(GroupPollVotes/ChatParticipants/RefreshTokens) -> count userId/tokenHash,
non-local return inside transaction lambda -> return@transaction, local
function forward reference in AiGatewayService (reordered runOnceThrottled),
call.remoteHost() helper usage, nullable-safe status, top-level const access.

+++ server/src/main/kotlin/com/maodouchat/server/repository/UserTagRepository.kt
@@ -89,8 +89,8 @@ class UserTagRepository {
                 .selectAll()
                 .where { UserTagAssignments.tagId inList tagIds }
                 .groupBy(UserTagAssignments.tagId)
-                .associate { it[UserTagAssignments.tagId] to it[UserTagAssignments.userId.count()].toInt() }
-        rows.map { row -> row.toTagRow(countByTag[row[UserTags.id]] ?: 0) }
+                .associate { it[UserTagAssignments.tagId] to it[UserTagAssignments.userId.count()] }
+        rows.map { row -> row.toTagRow(countByTag[row[UserTags.id]] ?: 0L) }
     }
 
     fun updateTag(
